+++ NEWS.md
@@ -3,6 +3,7 @@
 
  * Unify S3 method signatures to allow generic programming over them.
  * Generic record to record transform.
+ * Move more functions from DBI to rquery.
 
 # cdata 1.0.4 2019/01/04